chore: Refactor logparser validator
